Prepend '.' to reverse zone to force full octet match with PTR record
We have several /16 netblocks, e.g. 45.10.in-addr.arpa.  Inside of those /16s, some of the /24s are delegated, e.g. 3.45.10.in-addr.arpa, but many aren't.  

Because 'stripos($domain, $r["name"]);' allows for partial octet matching, occasionally automatically created PTR records were placed in the wrong zone.

For example, 5.123.45.10.in-addr.arpa is placed in 3.45.10.in-addr.arpa instead of 45.10.in-addr.arpa since we don't have the 123.45.10.in-addr.arpa zone delegated.

Prepending the reverse zone with a '.' prevents stripos from allowing partial octet matching.